# docs/data/swagger.json
{
  "basePath": "/",
  "consumes": [
    "application/json"
  ],
  "definitions": {
...
    "FeedbackModel": {
      "properties": {
        "answer": {
          "description": "The AI answer",
          "type": "string"
        },
        "api_key": {
          "description": "Optional API key",
          "type": "string"
        },
        "conversation_id": {
          "description": "id of the particular conversation",
          "type": "string"
        },
        "feedback": {
          "description": "User feedback",
          "type": "string"
        },
        "question": {
          "description": "The user question",
          "type": "string"
        },
        "question_index": {
          "description": "The question number in that particular conversation",
          "type": "integer"
        }
      },
      "required": [
        "conversation_id",
        "feedback",
        "question_index"
      ],
      "type": "object"
    },
...
    "ShareConversationModel": {
      "properties": {
        "chunks": {
          "description": "Chunks count (optional)",
          "type": "integer"
        },
        "conversation_id": {
          "description": "Conversation ID",
          "type": "string"
        },
        "prompt_id": {
          "description": "Prompt ID (optional)",
          "type": "string"
        },
        "user": {
          "description": "User ID (optional)",
          "type": "string"
        }
      },
      "required": [
        "conversation_id"
      ],
      "type": "object"
    },
...
  },
  "info": {
    "description": "API for DocsGPT",
    "title": "DocsGPT API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/feedback": {
      "post": {
        "description": "Submit feedback for a conversation",
        "operationId": "post_submit_feedback",
        "parameters": [
          {
            "in": "body",
            "name": "payload",
            "required": true,
            "schema": {
              "$ref": "#/definitions/FeedbackModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "conversations"
        ],
        "x-pat-scopes": [
          "conversations:write"
        ]
      }
    },
...
    "/api/share": {
      "post": {
        "description": "Share a conversation",
        "operationId": "post_share_conversation",
        "parameters": [
          {
            "in": "body",
            "name": "payload",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ShareConversationModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "sharing"
        ],
        "x-pat-denied": true
      }
    },
...
    "/api/shared_conversation/{identifier}": {
      "get": {
        "description": "Get publicly shared conversations by identifier",
        "operationId": "get_get_publicly_shared_conversations",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "sharing"
        ],
        "x-pat-denied": true
      },
      "parameters": [
        {
          "in": "path",
          "name": "identifier",
          "required": true,
          "type": "string"
        }
      ]
    },
...
  },
  "produces": [
    "application/json"
  ],
...
  "swagger": "2.0",
...
}